clean Modelモックを統一
fetch失敗時のモックとfetch成功時のモックを一つのモックに統一した。#refs 33
